// management/frontend/src/api/schemas/rolePublic.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Ego Collection Management
 * OpenAPI spec version: 0.1.0
 */

export interface RolePublic {
  id?: string;
  created_at?: string | null;
  updated_at?: string | null;
  deleted_at?: string | null;
  creator_id?: string | null;
  updater_id?: string | null;
  deleter_id?: string | null;
  /**
     * @minLength 1
     * @maxLength 64
     */
  role_name: string;
  business_line_id?: string | null;
  sort?: number;
  is_active?: boolean;
  remark?: string | null;
  workspace_id: string;
  business_line_name?: string | null;
  creator_name?: string | null;
  updater_name?: string | null;
}