// management/frontend/src/api/schemas/workspaceRolesReadRolesParams.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Ego Collection Management
 * OpenAPI spec version: 0.1.0
 */

export type WorkspaceRolesReadRolesParams = {
/**
 * @minimum 0
 */
skip?: number;
/**
 * @maximum 500
 */
limit?: number;
role_name?: string | null;
business_line_id?: string | null;
is_active?: boolean | null;
};