// management/frontend/src/api/zod/workspace-roles/workspace-roles.ts
export const workspaceRolesReadRolesQuerySkipDefault = 0;
export const workspaceRolesReadRolesQuerySkipMin = 0;

export const workspaceRolesReadRolesQueryLimitDefault = 20;
export const workspaceRolesReadRolesQueryLimitMax = 500;



export const WorkspaceRolesReadRolesQueryParams = zod.object({
  "skip": zod.number().min(workspaceRolesReadRolesQuerySkipMin).default(workspaceRolesReadRolesQuerySkipDefault),
  "limit": zod.number().max(workspaceRolesReadRolesQueryLimitMax).default(workspaceRolesReadRolesQueryLimitDefault),
  "role_name": zod.union([zod.string(),zod.null()]).optional(),
  "business_line_id": zod.union([zod.uuid(),zod.null()]).optional(),
  "is_active": zod.union([zod.boolean(),zod.null()]).optional()
})